fix(wos): four quick fixes for WOS orchestration layer
Fix 6: _DAN_CHAT_ID in steward.py now reads LOBSTER_ADMIN_CHAT_ID from the
environment with the hardcoded value as fallback, matching the pattern used
everywhere else in the codebase.

Fix 7: Split steward-heartbeat.py — startup sweep logic (Phase 1) extracted
into startup-sweep.py. steward-heartbeat.py imports and re-exports the sweep
symbols so existing tests that load the heartbeat via importlib continue to
work without modification.

Fix 8: Registry.get_started_at(uow_id) added as a public method. The
observation loop's _fetch_started_at() helper, which previously called
registry._connect() directly, is removed; the loop now calls the public API.

Fix 9: Verified all Phase 2 sqlite3.connect() calls (registry.py,
executor.py) already carry timeout=10.0 + PRAGMA busy_timeout=5000. No
changes needed.
